main into liam

+++ pid.h
@@ -17,8 +17,9 @@ extern float error;
 void pid_controller_loop(void);
 
 //PID data processing subroutines
-short derivative(void);
+float derivative(void);
 short integral(void);
 void circshift_single_insert(int* array, int element, int size);
+
 
 #endif
